+++ src/Sql/Sql.Test/ScenarioTests/AdvancedDataSecurityManagedInstanceTests.ps1
@@ -20,8 +20,8 @@ function Test-AdvancedDataSecurityPolicyManagedInstanceTest
 {
 	# Setup
 	$testSuffix = getAssetName
-	Create-AdvancedThreatProtectionManagedInstanceTestEnvironment $testSuffix
-	$params = Get-SqlAdvancedThreatProtectionManagedInstanceTestEnvironmentParameters $testSuffix
+	Create-AdvancedDataSecurityManagedInstanceTestEnvironment $testSuffix
+	$params = Get-SqlAdvancedDataSecurityManagedInstanceTestEnvironmentParameters $testSuffix
 
 	try
 	{
@@ -34,7 +34,7 @@ function Test-AdvancedDataSecurityPolicyManagedInstanceTest
 		Assert-False { $policy.IsEnabled }
 
 		# Enabled Advanced Threat Protection Policy
-		Enable-AzSqlInstanceAdvancedDataSecurity -ResourceGroupName $params.rgname -InstanceName $params.serverName 
+		Enable-AzSqlInstanceAdvancedDataSecurity -ResourceGroupName $params.rgname -InstanceName $params.serverName -DoNotConfigureVulnerabilityAssessment
 		$policy = Get-AzSqlInstanceAdvancedDataSecurityPolicy -ResourceGroupName $params.rgname -InstanceName $params.serverName 
 				
 		# Validate the policy
@@ -54,7 +54,7 @@ function Test-AdvancedDataSecurityPolicyManagedInstanceTest
 	finally
 	{
 		# Cleanup
-		Remove-AdvancedThreatProtectionManagedInstanceTestEnvironment $testSuffix
+		Remove-AdvancedDataSecurityManagedInstanceTestEnvironment $testSuffix
 	}
 }
 
@@ -62,9 +62,9 @@ function Test-AdvancedDataSecurityPolicyManagedInstanceTest
 .SYNOPSIS
 Creates the test environment needed to perform the tests
 #>
-function Create-AdvancedThreatProtectionManagedInstanceTestEnvironment ($testSuffix, $location = "West Central US")
+function Create-AdvancedDataSecurityManagedInstanceTestEnvironment ($testSuffix, $location = "West Central US")
 {
-	$params = Get-SqlAdvancedThreatProtectionManagedInstanceTestEnvironmentParameters $testSuffix
+	$params = Get-SqlAdvancedDataSecurityManagedInstanceTestEnvironmentParameters $testSuffix
 	Create-BasicManagedTestEnvironmentWithParams $params $location
 }
 
@@ -72,7 +72,7 @@ function Create-AdvancedThreatProtectionManagedInstanceTestEnvironment ($testSuf
 .SYNOPSIS
 Gets the values of the parameters used at the tests
 #>
-function Get-SqlAdvancedThreatProtectionManagedInstanceTestEnvironmentParameters ($testSuffix)
+function Get-SqlAdvancedDataSecurityManagedInstanceTestEnvironmentParameters ($testSuffix)
 {
 	return @{ rgname = "sql-atp-cmdlet-test-rg" +$testSuffix;
 			  serverName = "sql-atp-cmdlet-server" +$testSuffix;
@@ -84,8 +84,8 @@ function Get-SqlAdvancedThreatProtectionManagedInstanceTestEnvironmentParameters
 .SYNOPSIS
 Removes the test environment that was needed to perform the tests
 #>
-function Remove-AdvancedThreatProtectionManagedInstanceTestEnvironment ($testSuffix)
+function Remove-AdvancedDataSecurityManagedInstanceTestEnvironment ($testSuffix)
 {
-	$params = Get-SqlAdvancedThreatProtectionManagedInstanceTestEnvironmentParameters $testSuffix
+	$params = Get-SqlAdvancedDataSecurityManagedInstanceTestEnvironmentParameters $testSuffix
 	Remove-AzureRmResourceGroup -Name $params.rgname -Force
 }