+++ src/Websites/Websites.Test/ScenarioTests/WebAppBackupRestoreTests.cs
@@ -116,21 +116,21 @@ namespace Microsoft.Azure.Commands.Websites.Test.ScenarioTests
             WebsitesController.NewInstance.RunPsTest(_logger, "Test-RestoreWebAppSnapshot");
         }
 
-        [Fact]
+        [Fact(Skip="Get-AzDeletedWebApp returning null in recorded tests #9191")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestGetDeletedWebApp()
         {
             WebsitesController.NewInstance.RunPsTest(_logger, "Test-GetDeletedWebApp");
         }
 
-        [Fact]
+        [Fact(Skip="Get-AzDeletedWebApp returning null in recorded tests #9191")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestRestoreDeletedWebAppToExisting()
         {
             WebsitesController.NewInstance.RunPsTest(_logger, "Test-RestoreDeletedWebAppToExisting");
         }
 
-        [Fact]
+        [Fact(Skip="Get-AzDeletedWebApp returning null in recorded tests #9191")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestRestoreDeletedWebAppToNew()
         {