+++ src/Sql/Sql/ChangeLog.md
@@ -20,6 +20,7 @@
 ## Upcoming Release
 * Add DnsZonePartner Parameter for New-AzureSqlInstance cmdlet to support AutoDr for Managed Instance.
 * Rename Advanced Threat Protection cmdlets to Advanced Data Security and enable Vulnerability Assessment by default
+* Deprecating Get-AzSqlDatabaseSecureConnectionPolicy cmdlet
 
 ## Version 1.9.0
 * Replace dependency on Monitor SDK with common code